chore(watch): allows to watch files, not only dirs
