+++ cmd/ike/watch/watch_builder.go
@@ -1,0 +1,63 @@
+package watch
+
+import (
+	"os"
+
+	"github.com/fsnotify/fsnotify"
+)
+
+type Builder struct {
+	w *Watch
+}
+
+// CreateWatch creates instance of Builder providing fluent functions to customize watch
+func CreateWatch() *Builder {
+	return &Builder{w: &Watch{}}
+}
+
+// WithHandler allows to hook instance of Handler reacting on file change events
+func (wb *Builder) WithHandler(handler Handler) *Builder {
+	wb.w.handler = handler
+	return wb
+}
+
+// Excluding allows to define exclusion patterns (as glob expressions)
+func (wb *Builder) Excluding(exclusions ...string) *Builder {
+	wb.w.exclusions = ParseFilePatterns(exclusions)
+	return wb
+}
+
+// OnPaths defines paths to be watched.
+// If path is a directory it will recursively watch all files and subdirectories.
+// If path is a file only this path is watched
+func (wb *Builder) OnPaths(paths ...string) (watch *Watch, err error) {
+
+	fsWatcher, err := fsnotify.NewWatcher()
+	if err != nil {
+		log.Error(err, "failed creating fs watch")
+		return nil, err
+	}
+
+	wb.w.watcher = fsWatcher
+
+	for _, p := range paths {
+		dir, err := os.Stat(p)
+
+		if err != nil {
+			return nil, err
+		}
+
+		if !dir.IsDir() {
+			if e := wb.w.AddPath(p); e != nil {
+				return nil, e
+			}
+		} else {
+			if e := wb.w.AddRecursiveWatch(p); e != nil {
+				return nil, e
+			}
+		}
+
+	}
+
+	return wb.w, nil
+}
